changes made from theia lab 4 environment

+++ onlinecourse/models.py
@@ -96,7 +96,7 @@ class Enrollment(models.Model):
 class Question(models.Model):
     course = models.ForeignKey(Course, on_delete=models.CASCADE)
     content = models.CharField(max_length=200)
-    grade = models.IntegerField(default=50)
+    grade = models.FloatField(default=50)
 
     def __str__(self):
         return "Question: " + self.content
